requirements.txt
streamlit
torch
torchvision
transformers
pillow
requests
timm